third_party/blink/web_tests/shadow-dom/declarative-shadow-dom-errors-expected.txt

CONSOLE ERROR: This element does not support attachShadow
CONSOLE ERROR: attachShadow() is disabled by disabledFeatures static field.
CONSOLE ERROR: A second declarative shadow root cannot be created on a host.
PASS successfullyParsed is true

TEST COMPLETE
